refactor: modularize parallel_agent.py into agents/ package
Splits 2145-line monolith into 6 focused modules (config, validation,
synthesis, runners, orchestrator, cli) with a backward-compatible
entry-point shim. Adds 6 per-module unit test files (68 new tests),
updates existing test imports, and ratifies the project constitution.

128/128 tests pass. CLI interface unchanged.
