+++ specs/001-modularize-parallel-agent/baseline-test-results.txt
@@ -1,0 +1,60 @@
+============================= test session starts ==============================
+platform darwin -- Python 3.14.4, pytest-9.0.3, pluggy-1.6.0 -- /Library/Frameworks/Python.framework/Versions/3.14/bin/python3
+cachedir: .pytest_cache
+rootdir: /Users/chrismandich/Documents/GitHub/Manifest/.claude/worktrees/chore-refactor-parallel-agent
+plugins: anyio-4.13.0, asyncio-1.4.0
+asyncio: mode=Mode.STRICT, debug=False, asyncio_default_fixture_loop_scope=None, asyncio_default_test_loop_scope=function
+collecting ... collected 60 items
+
+tests/python/test_parallel_agent.py::TestConfig::test_default_config_when_file_missing PASSED [  1%]
+tests/python/test_parallel_agent.py::TestConfig::test_default_config_has_claude_tiers PASSED [  3%]
+tests/python/test_parallel_agent.py::TestConfig::test_default_config_has_gemini_tiers PASSED [  5%]
+tests/python/test_parallel_agent.py::TestConfig::test_get_dot_notation PASSED [  6%]
+tests/python/test_parallel_agent.py::TestConfig::test_get_returns_default_for_missing_key PASSED [  8%]
+tests/python/test_parallel_agent.py::TestConfig::test_get_returns_none_when_no_default PASSED [ 10%]
+tests/python/test_parallel_agent.py::TestConfig::test_load_from_yaml_file PASSED [ 11%]
+tests/python/test_parallel_agent.py::TestConfig::test_default_timeouts PASSED [ 13%]
+tests/python/test_parallel_agent.py::TestConfig::test_default_consensus_thresholds PASSED [ 15%]
+tests/python/test_parallel_agent.py::TestRateLimiter::test_creation_with_defaults PASSED [ 16%]
+tests/python/test_parallel_agent.py::TestRateLimiter::test_creation_with_custom_values PASSED [ 18%]
+tests/python/test_parallel_agent.py::TestRateLimiter::test_acquire_decrements_tokens PASSED [ 20%]
+tests/python/test_parallel_agent.py::TestRateLimiter::test_acquire_all_tokens PASSED [ 21%]
+tests/python/test_parallel_agent.py::TestRateLimiter::test_ignores_extra_kwargs PASSED [ 23%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_security_check_detects_hardcoded_secret PASSED [ 25%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_security_check_passes_clean_code PASSED [ 26%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_error_handling_detects_bare_except PASSED [ 28%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_bug_detection_scores_null_references PASSED [ 30%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_performance_check_detects_quadratic PASSED [ 31%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_verdict_approved PASSED [ 33%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_verdict_needs_review PASSED [ 35%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_verdict_blocked PASSED [ 36%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_verdict_with_custom_threshold_override PASSED [ 38%]
+tests/python/test_parallel_agent.py::TestValidationEngine::test_skips_incomplete_agents PASSED [ 40%]
+tests/python/test_parallel_agent.py::TestSynthesisEngine::test_creation PASSED [ 41%]
+tests/python/test_parallel_agent.py::TestSynthesisEngine::test_skips_synthesis_when_consensus_high PASSED [ 43%]
+tests/python/test_parallel_agent.py::TestSynthesisEngine::test_template_is_string PASSED [ 45%]
+tests/python/test_parallel_agent.py::TestBaseAgent::test_credit_exhaustion_detection_quota PASSED [ 46%]
+tests/python/test_parallel_agent.py::TestBaseAgent::test_credit_exhaustion_detection_rate_limit PASSED [ 48%]
+tests/python/test_parallel_agent.py::TestBaseAgent::test_credit_exhaustion_detection_429 PASSED [ 50%]
+tests/python/test_parallel_agent.py::TestBaseAgent::test_no_false_positive_on_normal_error PASSED [ 51%]
+tests/python/test_parallel_agent.py::TestBaseAgent::test_execute_returns_result PASSED [ 53%]
+tests/python/test_parallel_agent.py::TestBaseAgent::test_execute_timeout PASSED [ 55%]
+tests/python/test_parallel_agent.py::TestOrchestratorConsensus::test_consensus_with_similar_outputs PASSED [ 56%]
+tests/python/test_parallel_agent.py::TestOrchestratorConsensus::test_consensus_with_single_agent PASSED [ 58%]
+tests/python/test_parallel_agent.py::TestOrchestratorConsensus::test_consensus_with_no_complete_agents PASSED [ 60%]
+tests/python/test_parallel_agent.py::TestOrchestratorConsensus::test_confidence_levels PASSED [ 61%]
+tests/python/test_parallel_agent.py::TestLogger::test_logger_creation PASSED [ 63%]
+tests/python/test_parallel_agent.py::TestLogger::test_correlation_id PASSED [ 65%]
+tests/python/test_parallel_agent.py::TestLogger::test_logging_methods_do_not_raise PASSED [ 66%]
+tests/python/test_parallel_agent.py::TestJSONOutputFormat::test_mock_fixture_matches_schema PASSED [ 68%]
+tests/python/test_parallel_agent.py::TestServiceConfig::test_defaults_when_file_missing PASSED [ 70%]
+tests/python/test_parallel_agent.py::TestServiceConfig::test_minimum_agents_default PASSED [ 71%]
+tests/python/test_parallel_agent.py::TestServiceConfig::test_is_enabled_from_yaml PASSED [ 73%]
+tests/python/test_parallel_agent.py::TestServiceConfig::test_check_minimum_agents_ok PASSED [ 75%]
+tests/python/test_parallel_agent.py::TestServiceConfig::test_check_minimum_agents_warning PASSED [ 76%]
+tests/python/test_parallel_agent.py::TestServiceConfig::test_unknown_service_defaults_enabled PASSED [ 78%]
+tests/python/test_parallel_agent.py::TestCodexAgent::test_resolve_model_auto PASSED [ 80%]
+tests/python/test_parallel_agent.py::TestCodexAgent::test_resolve_model_named_tier PASSED [ 81%]
+tests/python/test_parallel_agent.py::TestCodexAgent::test_resolve_model_custom PASSED [ 83%]
+tests/python/test_parallel_agent.py::TestCodexAgent::test_execute_missing_codex PASSED [ 85%]
+tests/python/test_parallel_agent.py::TestDefaultConfigCodex::test_default_config_has_codex_rate_limits PASSED [ 86%]
