+++ tests/python/agents/conftest.py
@@ -1,0 +1,9 @@
+"""Pytest configuration — adds agents package to sys.path for per-module tests."""
+import sys
+from pathlib import Path
+
+REPO_ROOT = Path(__file__).resolve().parent.parent.parent.parent
+SCRIPTS_DIR = str(REPO_ROOT / "configs" / "claude" / "scripts")
+
+if SCRIPTS_DIR not in sys.path:
+    sys.path.insert(0, SCRIPTS_DIR)
